Rework Ansible and ssh key handling
The AnsibleRunner should be a part of the Workspace as it is sometimes
used to execute tasks on the local machine.

Also move the ssh-agent into the Workspace as it is not only needed for
CaaSP.

To simplify the ssh-key management, also move these into the Workspace.
Arguably we could have a different private-key per Hardware, but having
one generated for the workspace should be sufficient.

This also fixes the modification of env vars for setting the ssh-agent.
