+++ tests/lib/hardware/openstack_libcloud.py
@@ -203,16 +203,13 @@ class Hardware(HardwareBase):
     def __init__(self, workspace: Workspace):
         super().__init__(workspace)
         self._ex_os_key = self.conn.import_key_pair_from_string(
-            self.sshkey_name, self.public_key)
+            self.workspace.sshkey_name, self.workspace.public_key)
         self._ex_security_group: OpenStackSecurityGroup = \
             self._create_security_group()
         self._ex_network_cache: List[OpenStackNetwork] = []
 
         self._image_cache: Dict[str, NodeImage] = {}
         self._size_cache: List[OpenStackNodeSize] = []
-
-        logger.info(f"public key {self.public_key}")
-        logger.info(f"private key {self.private_key}")
 
     def get_connection(self):
         """ Get a libcloud connection object for the configured driver """
@@ -319,7 +316,7 @@ class Hardware(HardwareBase):
                     self._get_size_by_name(config.NODE_SIZE),
                     self._get_image(config.NODE_IMAGE),
                     additional_networks, [self._ex_security_group],
-                    self.sshkey_name)
+                    self.workspace.sshkey_name)
 
         node.boot()
         self.node_add(node)
