+++ tests/lib/kubernetes/caasp.py
@@ -35,30 +35,6 @@ class CaaSP(KubernetesBase):
         self._clusterpath = os.path.join(self.workspace.working_dir, 'cluster')
         self._kubeconfig = os.path.join(self.workspace.working_dir,
                                         'admin.conf')
-        self._ssh_agent_auth_sock = os.path.join(hardware.working_dir,
-                                                 'ssh-agent.sock')
-        self._ssh_agent()
-
-    def _ssh_agent(self):
-        try:
-            res = subprocess.run(
-                ['ssh-agent', '-a', self._ssh_agent_auth_sock],
-                check=True, capture_output=True)
-        except subprocess.CalledProcessError:
-            logger.exception('Failed to start ssh agent')
-            raise
-
-        logger.info(f"ssh-agent started: {res.stdout.decode('utf-8')}")
-        self._ssh_agent_pid = res.stdout.decode(
-            'utf-8').split(';')[2].split('=')[1]
-        os.environ['SSH_AUTH_SOCK'] = self._ssh_agent_auth_sock
-        os.environ['SSH_AGENT_PID'] = self._ssh_agent_pid
-        try:
-            res = subprocess.run(['ssh-add', self.hardware._private_key],
-                                 check=True)
-        except subprocess.CalledProcessError:
-            logger.exception('Failed to add keys to agent')
-            raise
 
     def destroy(self, skip=False):
         logger.info(f"kube destroy on hardware {self.hardware}")
@@ -66,13 +42,6 @@ class CaaSP(KubernetesBase):
             # We can skip in most cases since the nodes themselves will be
             # destroyed instead.
             return
-        # This kills the SSH_AGENT_PID agent
-        try:
-            subprocess.run(['ssh-agent', '-k'], check=True)
-        except subprocess.CalledProcessError:
-            logger.exception(f'Killing ssh-agent with PID \
-{self._ssh_agent_pid} failed')
-
         # TODO(jhesketh): Uninstall kubernetes
 
     def install_kubernetes(self):
